fix broken links, move demoURL to constants file
