+++ apps/docs-website/docs/getting-started-with-loon/trackmate-data.md
@@ -1,13 +1,15 @@
 # Optional: Use Data from TrackMate
 
 #### [TrackMate](https://imagej.net/plugins/trackmate/) is a popular FIJI (Image Analysis) plugin for:
-  - Robust tracking of cells and other objects in microscopy time-lapse images
-  - Various detection and tracking algorithms
-  - Visualization and analysis tools
-  - [Cellpose](https://www.cellpose.org/) is a cell segmentation algorithm for TrackMate:
-    - Enables advanced object detection directly within the TrackMate workflow
+
+- Robust tracking of cells and other objects in microscopy time-lapse images
+- Various detection and tracking algorithms
+- Visualization and analysis tools
+- [Cellpose](https://www.cellpose.org/) is a cell segmentation algorithm for TrackMate:
+  - Enables advanced object detection directly within the TrackMate workflow
 
 ## 1:&nbsp;&nbsp;&nbsp;Export Data from [TrackMate](https://imagej.net/plugins/trackmate/)
+
 > ### 1.1:&nbsp;&nbsp;Rename tracks in Trackmate for lineage tracking
 >
 > - After a cell division, identify the parent track and the two daughter tracks.
@@ -25,26 +27,33 @@
 > - Export ImageJ ROIs: This is an important step to get the segmented regions of cells themselves. This can be done in ImageJ after using TrackMate to generate the ROIs, and typically involves saving the ROI Manager contents as a .zip file.
 
 ## 2:&nbsp;&nbsp;&nbsp;Convert TrackMate Data into [Loon Format](./data.md)
+
 > We currently use [our conversion script on github](https://github.com/visdesignlab/aardvark-util/blob/main/ingest_trackmate.py). Click the Download button at the top-right of the script.
+>
 > ### Conversion Script Info:
->> **Inputs:**
->> - A `.csv` file from TrackMate  
->>   - Must currently include ``LABEL``, ``FRAME``, ``POSITION_X``, ``POSITION_Y`` columns  
->> - A folder containing `.roi` files from TrackMate
->> 
->> **What the script does:**  
->> - Reads your `.csv` file, removes unnecessary rows / columns, sorts by frame  
->> - Infers / Adds a `parent` column to the `.csv` file, which includes the parents of each track  
->> - Outputs that corrected `.csv`  
->> - Converts that corrected `.csv` to a `.parquet` file  
->> - Converts `.roi` files to `GeoJSON` format, creating a folder structure based on frames  
->> 
->> **Outputs:**  
->> - A `metadata.csv` file with metadata for Loon  
->> - A `metadata.parquet` file with metadata for Loon  
->> - A segmentations folder with `GeoJSON` files for each frame  
->>
+>
+> > **Inputs:**
+> >
+> > - A `.csv` file from TrackMate
+> >   - Must currently include `LABEL`, `FRAME`, `POSITION_X`, `POSITION_Y` columns
+> > - A folder containing `.roi` files from TrackMate
+> >
+> > **What the script does:**
+> >
+> > - Reads your `.csv` file, removes unnecessary rows / columns, sorts by frame
+> > - Infers / Adds a `parent` column to the `.csv` file, which includes the parents of each track
+> > - Outputs that corrected `.csv`
+> > - Converts that corrected `.csv` to a `.parquet` file
+> > - Converts `.roi` files to `GeoJSON` format, creating a folder structure based on frames
+> >
+> > **Outputs:**
+> >
+> > - A `metadata.csv` file with metadata for Loon
+> > - A `metadata.parquet` file with metadata for Loon
+> > - A segmentations folder with `GeoJSON` files for each frame
+>
 > ### Run the script:
+>
 > 1. Open your terminal:
 >    - **Mac:** Press `Cmd + Space`, type `Terminal`, and press `Enter`
 >    - **Windows:** Press `Win + R`, type `cmd`, and press `Enter`
@@ -70,10 +79,10 @@
 > 4. Follow the on-screen instructions to convert your TrackMate data to Loon format
 > </details>
 
-
-## 3:&nbsp;&nbsp;&nbsp;Upload Formatted Data to Loon
-
-> ### Option 1: [Local Loon Upload](/.quickstart.md)
-> ### Option 2: [Server Upload](/.loon-wrappers.md)
+<!-- TODO: point to something that exists -->
+<!-- ## 3:&nbsp;&nbsp;&nbsp;Upload Formatted Data to Loon
 
 
+> ### See: [Local Loon Upload](./quickstart.md)
+>
+> ### Option 2: [Server Upload](./loon-wrappers.md) -->
